Fixed simple modes issue: Sorting a table triggered delete action

+++ cmk/gui/plugins/wato/utils/simple_modes.py
@@ -174,6 +174,9 @@ class SimpleListMode(SimpleWatoModeBase):
         if not html.transaction_valid():
             return
 
+        if not html.has_var("_delete"):
+            return
+
         confirm = wato_confirm(_("Confirm deletion"), self._delete_confirm_message())
         if confirm is False:
             return False
